load from a dashboard_state.json
