+++ src/hypofuzz/frontend/src/App.tsx
@@ -4,8 +4,10 @@ import { TestsPage } from "./pages/Tests"
 import { PatchesPage } from "./pages/Patches"
 import { CollectionStatusPage } from "./pages/CollectionStatus"
 import { DevPage } from "./pages/Dev"
-import { WebSocketProvider } from "./context/WebSocketContext"
+import { NotFoundPage } from "./pages/NotFound"
+import { DataProvider } from "./context/DataProvider"
 import { Layout } from "./components/Layout"
+import React from "react"
 
 export function App() {
   return (
@@ -15,15 +17,16 @@ export function App() {
           <Route
             path="/"
             element={
-              <WebSocketProvider>
+              <DataProvider>
                 <TestsPage />
-              </WebSocketProvider>
+              </DataProvider>
             }
           />
           <Route path="/patches" element={<PatchesPage />} />
           <Route path="/collected" element={<CollectionStatusPage />} />
           <Route path="/tests/:testId" element={<TestPage />} />
           <Route path="/_dev" element={<DevPage />} />
+          <Route path="*" element={<NotFoundPage />} />
         </Route>
       </Routes>
     </BrowserRouter>
